requirements.txt
lxml
PyPDF2
python-dotenv
google-generativeai
pinecone-client
fuzzywuzzy
python-Levenshtein
PyMuPDF
